+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 ---
 
 - [feat:support setting load balancing strategy per service.](https://github.com/Tencent/spring-cloud-tencent/pull/1633)
+- [feat: support shortest response time lb and least connection lb](https://github.com/Tencent/spring-cloud-tencent/pull/1636)